requirements.txt
Django>=1.5
django-model-utils
factory_boy>=1.3,<2.0